% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_local_repos.R
\name{fix_bioc_remotes}
\alias{fix_bioc_remotes}
\title{A helper function to adjust git remotes the Bioconductor way}
\usage{
fix_bioc_remotes(repo_dir, org = "Bioconductor", is_bioc)
}
\arguments{
\item{repo_dir}{\code{character(1)} The full path to the directory of the local
repository whose default branch should be updated}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{is_bioc}{\code{logical(1)} Whether the repository in question is a
Bioconductor package}
}
\value{
Called for the side-effect of adjusting \verb{git remotes} in the
\code{repo_dir}; check them with \verb{git remote -v}
}
\description{
This function will update a repository's remotes by setting \code{origin} to the
GitHub location and \code{upstream} to the Bioconductor git server.
}
